+++ docutils/HISTORY.rst
@@ -32,10 +32,18 @@ Release 0.23b0 (unpublished)
   - Ensure new "current node" is valid when switching section level
     (cf. bugs #508 and #509).
   - Use a `separate title style hierarchy for nested parsing`__.
+  - Set `parent_state_machine` attribute when creating nested state machines.
+    Allows passing an updated "current node" to the parent state machine,
+    e.g. for changing the section level in a directive.
   - Better error messages for grid table markup errors (bug #504),
     based on patch #214 by Jynn Nelson.
 
   __ RELEASE-NOTES.html#nested-parsing
+
+* docutils/statemachine.py
+
+  - New attribute `StateMachine.parent_state_machine` to store the
+    parent state machine of nested state machines.
 
 * docutils/transforms/references.py